requirements.txt
adlfs
azure-storage-blob
azureml-mlflow
datasets
lightning
mlflow
onnx
peft
polars
sentencepiece
torch
torchmetrics
transformers
pandas
matplotlib
hydra-core
pydantic
google_cloud_mlflow